requirements.txt
funcsigs==1.0.2
mock==2.0.0
numpy==1.13.3
pbr==3.1.1
pkg-resources==0.0.0
protobuf==3.5.0.post1
scipy==0.19.1
six==1.11.0
tensorflow-gpu==0.12.1
